Seção 5 - Desenvolvimento de API's Restful (Finalizando implementações e validações seguindo o contrato da API)

+++ libraryapi/src/main/java/br/com/murilo/libraryapi/service/AutorService.java
@@ -1,21 +1,43 @@
 package br.com.murilo.libraryapi.service;
 
+import br.com.murilo.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import br.com.murilo.libraryapi.model.Autor;
 import br.com.murilo.libraryapi.repository.AutorRepository;
+import br.com.murilo.libraryapi.repository.LivroRepository;
+import br.com.murilo.libraryapi.validator.AutorValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class AutorService {
 
-    @Autowired
-    private AutorRepository autorRepository;
+    // camada de serviço - parte lógica
+
+    private final AutorRepository autorRepository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
+
+    public AutorService(AutorRepository autorRepository, AutorValidator validator, LivroRepository livroRepository) {
+        this.autorRepository = autorRepository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
+    }
 
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return autorRepository.save(autor);
+    }
+
+    public void atualizar(Autor autor) {
+        if (autor.getId() == null) {
+            throw new IllegalArgumentException("Para atualizar é necessário que o autor esteja salvo na base.");
+        }
+        validator.validar(autor);
+        autorRepository.save(autor);
     }
 
     public Optional<Autor> obterPorId(UUID id) {
@@ -23,8 +45,31 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor que possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
     }
 
+    // não é a melhor prática
+    public List<Autor> pesquisa(String nome, String nacionalidade) {
+        if (nome != null && nacionalidade != null) {
+            return autorRepository.findByNomeAndNacionalidade(nome, nacionalidade);
+        }
+
+        if (nome != null) {
+            return autorRepository.findByNome(nome);
+        }
+
+        if (nacionalidade != null) {
+            return autorRepository.findByNacionalidade(nacionalidade);
+        }
+
+        return autorRepository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
+    }
 
 }
